+++ Core/TimberApi/ToolSystem/ToolService.cs
@@ -8,7 +8,7 @@ using Timberborn.ToolSystem;
 
 namespace TimberApi.ToolSystem
 {
-    public class ToolService : IObjectSpecificationLoadableSingleton
+    public class ToolService : ILateLoadableSingleton
     {
         private readonly ToolButtonFactoryService _toolButtonFactoryService;
 
@@ -34,13 +34,12 @@ namespace TimberApi.ToolSystem
 
         public IEnumerable<ToolButton> ToolButtons => _toolButtons.Select(pair => pair.Value).ToImmutableArray();
 
-        public void SpecificationLoad()
+        public void LateLoad()
         {
             var tools = new Dictionary<string, Tool>();
 
             var toolButtons = new Dictionary<string, ToolButton>();
-
-
+            
             foreach (var specification in _toolSpecificationService.ToolSpecifications)
             {
                 var toolFactory = _toolFactoryService.Get(specification.Type);