+++ Core/TimberApi/ToolSystem/ToolSpecificationService.cs
@@ -4,11 +4,10 @@ using System.Collections.Immutable;
 using System.Linq;
 using TimberApi.Common.SingletonSystem;
 using Timberborn.Persistence;
-using UnityEngine;
 
 namespace TimberApi.ToolSystem
 {
-    public class ToolSpecificationService : IObjectSpecificationLoadableSingleton
+    public class ToolSpecificationService : ILateLoadableSingleton
     {
         private readonly ISpecificationService _specificationService;
 
@@ -26,7 +25,7 @@ namespace TimberApi.ToolSystem
 
         public ImmutableArray<ToolSpecification> ToolSpecifications => _toolSpecifications.Select(pair => pair.Value).ToImmutableArray();
 
-        public void SpecificationLoad()
+        public void LateLoad()
         {
             _toolSpecifications = _specificationService.GetSpecifications(_toolSpecificationDeserializer).ToImmutableDictionary(specification => specification.Id.ToLower());
         }